+++ surya/scripts/__init__.py
@@ -1,0 +1,5 @@
+from surya.scripts.detect_layout import detect_layout_cli
+from surya.scripts.detect_text import detect_text_cli
+from surya.scripts.run_streamlit_app import streamlit_app_cli
+from surya.scripts.ocr_text import ocr_text_cli
+from surya.scripts.table_recognition import table_recognition_cli
